FastApiServer/requirements.txt
Fastapi
uvicorn
python-dotenv
datetime
tinkoff-investments
apscheduler
schedule

